empezando devolucion, fix sqlcon dao, agregado selectall

+++ src/AerolineaFrba/Models/Pasaje.cs
@@ -10,8 +10,8 @@ namespace AerolineaFrba.Models {
     class Pasaje {
 
         public int Id { get; set; }
-        public int Codigo { get; set; }
-        public float Precio { get; set; }
+        public Decimal Codigo { get; set; }
+        public Decimal Precio { get; set; }
         public DateTime Fecha_Compra { get; set; }
 
         public int Butaca_Id { get; set; }
